rename folder "pack" to "standalone"
