+++ lib/standalone/index.ts
